Remove ManagedAccountMain
